Add options to allow DropUp menu's.
Added Styling for DropUp.
Added options for DropUp, and override if Toolbar is positioned at bottom.
However, in the Paragraph function in `src/js/base/module/Buttons.js` where the function resides near line 520, I could not work out how to add the option `dropUp: (this.options.toolbarPosition === 'bottom' ? (this.options.toolbarBottomDropUp ? true : this.options.dropUp ) : this.options.dropUp),` so if the Dropdown is supposed to be changed to DropUp.