Add option to allow FileBrowser Button to appear
Added option, markup and functionality to allow a FileBrowser button to appear in ImageDialog.js, it will replace the Browse Files input in favour of a File Browser Addon that does the same thing. Only the ID is parsed, and the user needs to create the function to open the File Browser and insert the returned File URL, which gets places into the URL Input.