Add Extra Attributes for ImageDialog.js and some other fixes.
Add attributes to ImageDialog.js, still needs work as I would like it to be able to behave like the LinkDialog where it reads the attributes for editing.
Fixed the control area when images are selected, the z-index needed to be increased.
Removed the extra unnecessary style from dialogs that could cause unwanted style issues when integrated into projects.