Fullscreen and Statusbar mods
I noticed when in Fullscreen that the editing area is transparent rather than solid, the var() colour was not being used.
Changed the behaviour of the resize function to use the actual resize markup area, so when in Fullscreen the editor can't be resized, and added a value so the editor would take into consideration the status bar so it sits inside the window area rather than making the editor higher pushing the statusbar off canvas.
Changed the default height of the output status bar area so it's always visible, eliminating height changes when status output is used.