Allow click on Modal Backdrop to close modal
Removed adding the element for backdrop as I noticed it being added everytime a modal is open (every time), and not being removed. So I moved the styling to still have the darkened background under a modal, which allowed for binding an event for closing the modal when clicking the backdrop. This reduced code, and stopped spawning of unnused backdrop elements.