Modify Dropdzone.js
As per Issue: https://github.com/summernote/summernote/issues/3247

There was also an issue reported that dropzone wasn't disabled, so I modified the check and move the $eventListener into the if false statement area so they are not attached to the DOM if dropzone is enabled. Dropping defaults back to browser handling.